requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiohttp_socks==0.8.0
dhooks==1.1.4
fastapi==0.101.0
httpx==0.24.1
motor==3.2.0
openai==0.27.8
providers==0.0.2
pydantic==2.1.1
pymongo==4.4.1
python-dotenv==1.0.0
python_socks==2.3.0
pytz==2023.3
PyYAML==6.0.1
rich==13.5.2
setuptools==65.5.1
starlette
tiktoken==0.4.0
